S08P22A703-219 feat : 기부된 목소리 조회 기능 구현

+++ backend/src/main/java/com/soundbridge/domain/voice/request/VoiceListConditionReq.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 public class VoiceListConditionReq {
     private Integer voiceAge;
     private String voiceGender;
-    private List<Integer> features;
+    private List<Long> features;
 
     @Override
     public String toString() {
